src/main/java/de/innovationhub/prox/modules/project/application/project/usecase/commands/StateInterestHandler.java
package de.innovationhub.prox.modules.project.application.project.usecase.commands;

import de.innovationhub.prox.modules.commons.application.ApplicationComponent;
import de.innovationhub.prox.modules.project.application.project.exception.ProjectNotFoundException;
import de.innovationhub.prox.modules.project.domain.project.Project;
import de.innovationhub.prox.modules.project.domain.project.ProjectRepository;
import java.util.UUID;
import lombok.RequiredArgsConstructor;

@ApplicationComponent
@RequiredArgsConstructor
public class StateInterestHandler {
  private final ProjectRepository projectRepository;

  public Project handle(UUID projectId, UUID userId) {
    var project = projectRepository.findById(projectId)
        .orElseThrow(ProjectNotFoundException::new);

    project.stateInterest(userId);

    return projectRepository.save(project);
  }
}
